+++ src/components/KataPage.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import Particles from 'react-particles-js';
-import pParams from '../particles/particles';
+// import Particles from 'react-particles-js';
+// import pParams from '../particles/particles';
 import '../css/style.css'
 import CircularProgressbar from 'react-circular-progressbar';
